rename viewcollections to collectionsgrid
