+++ src/main/kotlin/org/wycliffeassociates/otter/jvm/app/ui/mainscreen/viewmodel/MainViewViewModel.kt
@@ -5,7 +5,7 @@ import javafx.beans.property.SimpleStringProperty
 import org.wycliffeassociates.otter.common.data.model.Collection
 import org.wycliffeassociates.otter.common.data.model.Content
 import org.wycliffeassociates.otter.jvm.app.ui.mainscreen.view.MainScreenView
-import org.wycliffeassociates.otter.jvm.app.ui.viewcollections.view.ViewCollections
+import org.wycliffeassociates.otter.jvm.app.ui.collectionsgrid.view.CollectionsGrid
 import org.wycliffeassociates.otter.jvm.app.ui.viewcontent.view.ViewContent
 import org.wycliffeassociates.otter.jvm.app.ui.viewtakes.view.ViewTakesView
 import tornadofx.*
@@ -46,8 +46,8 @@ class MainViewViewModel: ViewModel() {
     fun projectSelected(selectedProject: Collection) {
         setActiveProjectText(selectedProject)
 
-        find<MainScreenView>().activeFragment.dock<ViewCollections>()
-        ViewCollections().apply {
+        find<MainScreenView>().activeFragment.dock<CollectionsGrid>()
+        CollectionsGrid().apply {
             activeProject.bindBidirectional(selectedProjectProperty)
             activeCollection.bindBidirectional(selectedCollectionProperty)
         }
