+++ src/main/kotlin/org/wycliffeassociates/otter/jvm/app/ui/viewtakes/view/ViewTakesStyles.kt
@@ -5,7 +5,6 @@ import javafx.scene.Cursor
 import javafx.scene.effect.DropShadow
 import javafx.scene.paint.Color
 import org.wycliffeassociates.otter.jvm.app.theme.AppTheme
-import org.wycliffeassociates.otter.jvm.app.ui.projecthome.view.ProjectHomeStyles
 import tornadofx.Stylesheet
 import tornadofx.box
 import tornadofx.cssclass
